Add files via upload
Logarithmic regression is a type of regression analysis, a statistical method used in various fields, including finance and stock trading. By definition, if the dependent variable increases or decreases (e.g. stock price) rapidly at first as the independent variable decreases (e.g. time), but the rate of that increase or decrease gradually slows, just like how stock price behaves, you should consider using a logarithmic regression.

Logarithmic regression is often used to predict trends and to discover patterns more easily than drawing a straight line on a linear stock chart.

The general form of a logarithmic regression equation is:

ln(y) = α+β ln (x)

We take the natural log (ln) on y here as the residuals (here, the stock prices) are not normally distributed, taking the logarithm of a skewed variable may improve the fit by altering the scale and making the variable more “normally” distributed.

Taking log also will “straighten” the indicator and force the stock price (or the data) to fit into the log regression area in a linear fashion for better visibility/presentation.